+++ app/models/atlas_engine/address_validation/validators/full_address/address_comparison.rb
@@ -9,17 +9,13 @@ module AtlasEngine
           extend T::Sig
           include Comparable
 
+          attr_reader :address, :candidate, :datastore
+
           sig { params(address: AbstractAddress, candidate: Candidate, datastore: DatastoreBase).void }
           def initialize(address:, candidate:, datastore:)
-            @street_comparison = StreetComparison.new(address: address, candidate: candidate, datastore: datastore)
-            @city_comparison = CityComparison.new(address: address, candidate: candidate, datastore: datastore)
-            @province_code_comparison = ProvinceCodeComparison.new(
-              address: address,
-              candidate: candidate,
-              datastore: datastore,
-            )
-            @zip_comparison = ZipComparison.new(address: address, candidate: candidate, datastore: datastore)
-            @building_comparison = BuildingComparison.new(address: address, candidate: candidate, datastore: datastore)
+            @address = address
+            @candidate = candidate
+            @datastore = datastore
           end
 
           sig { params(other: AddressComparison).returns(Integer) }
@@ -41,32 +37,40 @@ module AtlasEngine
 
           sig { returns(T::Boolean) }
           def potential_match?
-            street_comparison.nil? || T.must(street_comparison).potential_match?
+            street_comparison.sequence_comparison.nil? || T.must(street_comparison.sequence_comparison).potential_match?
           end
 
-          sig { returns(T.nilable(Token::Sequence::Comparison)) }
+          sig { returns(ZipComparison) }
           def zip_comparison
-            @zip_comparison.compare
+            @zip_comparison ||= ZipComparison.new(address: address, candidate: candidate, datastore: datastore)
           end
 
-          sig { returns(T.nilable(Token::Sequence::Comparison)) }
+          sig { returns(StreetComparison) }
           def street_comparison
-            @street_comparison.compare
+            @street_comparison ||= StreetComparison.new(address: address, candidate: candidate, datastore: datastore)
           end
 
-          sig { returns(T.nilable(Token::Sequence::Comparison)) }
+          sig { returns(CityComparison) }
           def city_comparison
-            @city_comparison.compare
+            @city_comparison ||= CityComparison.new(address: address, candidate: candidate, datastore: datastore)
           end
 
-          sig { returns(T.nilable(Token::Sequence::Comparison)) }
+          sig { returns(ProvinceCodeComparison) }
           def province_code_comparison
-            @province_code_comparison.compare
+            @province_code_comparison ||= ProvinceCodeComparison.new(
+              address: address,
+              candidate: candidate,
+              datastore: datastore,
+            )
           end
 
-          sig { returns(NumberComparison) }
+          sig { returns(BuildingComparison) }
           def building_comparison
-            @building_comparison.compare
+            @building_comparison ||= BuildingComparison.new(
+              address: address,
+              candidate: candidate,
+              datastore: datastore,
+            )
           end
 
           protected
@@ -76,21 +80,21 @@ module AtlasEngine
           end
           def comparisons
             [
-              street_comparison,
-              city_comparison,
-              zip_comparison,
-              province_code_comparison,
-              building_comparison,
+              street_comparison.sequence_comparison,
+              city_comparison.sequence_comparison,
+              zip_comparison.sequence_comparison,
+              province_code_comparison.sequence_comparison,
+              building_comparison.sequence_comparison,
             ].compact_blank
           end
 
           sig { returns(T::Array[AtlasEngine::AddressValidation::Token::Sequence::Comparison]) }
           def text_comparisons
             [
-              street_comparison,
-              city_comparison,
-              zip_comparison,
-              province_code_comparison,
+              street_comparison.sequence_comparison,
+              city_comparison.sequence_comparison,
+              zip_comparison.sequence_comparison,
+              province_code_comparison.sequence_comparison,
             ].compact_blank
           end