+++ test/countries/atlas_engine/si/address_validation/exclusions/unknown_city_test.rb
@@ -56,10 +56,13 @@ module AtlasEngine
           def mock_address_comparison(given_city, candidate_city)
             given_sequence = AtlasEngine::AddressValidation::Token::Sequence.from_string(given_city)
             candidate_sequence = AtlasEngine::AddressValidation::Token::Sequence.from_string(candidate_city)
-            city_comparison = AtlasEngine::AddressValidation::Token::Sequence::Comparator.new(
+            city_comparison_result = AtlasEngine::AddressValidation::Token::Sequence::Comparator.new(
               left_sequence: given_sequence,
               right_sequence: candidate_sequence,
             ).compare
+
+            city_comparison = typed_mock(AtlasEngine::AddressValidation::Validators::FullAddress::CityComparison)
+            city_comparison.stubs(:sequence_comparison).returns(city_comparison_result)
 
             address_comparison = typed_mock(AtlasEngine::AddressValidation::Validators::FullAddress::AddressComparison)
             address_comparison.stubs(:city_comparison).returns(city_comparison)