+++ test/models/atlas_engine/address_validation/validators/full_address/zip_comparison_test.rb
@@ -13,14 +13,14 @@ module AtlasEngine
           include AddressValidation::TokenHelper
           include AddressValidationTestHelper
 
-          test "#compare compares the session zip with the candidate zip field" do
+          test "#sequence_comparison compares the session zip with the candidate zip field" do
             candidate = Candidate.new(id: "A", source: { "zip" => "J9A 2V2" })
             address = build_address(country_code: "CA", zip: "j9a2v2")
             datastore = Es::Datastore.new(address: address)
 
             zip_comparison = ZipComparison.new(address:, candidate:, datastore:)
 
-            comparison = zip_comparison.compare
+            comparison = zip_comparison.sequence_comparison
             candidate_zip_sequences = candidate.component(:zip).sequences
 
             assert_predicate comparison, :match?
@@ -29,14 +29,14 @@ module AtlasEngine
             assert_equal candidate_zip_sequences.first, comparison.right_sequence
           end
 
-          test "#compare compares the session zip with a truncated candidate zip field when applicable" do
+          test "#sequence_comparison compares the session zip with a truncated candidate zip field when applicable" do
             candidate = Candidate.new(id: "A", source: { "zip" => "S2919 BNA" })
             address = build_address(country_code: "AR", zip: "S2919")
             datastore = Es::Datastore.new(address: address)
 
             zip_comparison = ZipComparison.new(address:, candidate:, datastore:)
 
-            comparison = zip_comparison.compare
+            comparison = zip_comparison.sequence_comparison
 
             candidate.component(:zip).value = "S2919"
             expected_candidate_zip_sequences = candidate.component(:zip).sequences
@@ -47,14 +47,14 @@ module AtlasEngine
             assert_equal expected_candidate_zip_sequences.first, comparison.right_sequence
           end
 
-          test "#compare returns nil comparison for candidate when there is no field value to compare" do
+          test "#sequence_comparison returns nil comparison for candidate when there is no field value to compare" do
             candidate = Candidate.new(id: "A", source: { "zip" => nil })
             address = build_address(country_code: "CA", zip: "j9a2v2")
             datastore = Es::Datastore.new(address: address)
 
             zip_comparison = ZipComparison.new(address:, candidate:, datastore:)
 
-            comparison = zip_comparison.compare
+            comparison = zip_comparison.sequence_comparison
             candidate_zip_sequences = candidate.component(:zip).sequences
 
             assert_nil comparison